ci: 拆分 sanitizer 为 ASAN/TSAN 独立模式，补全覆盖盲区
此前 TASKFLOWLITE_ENABLE_SANITIZER 是一个布尔开关，按操作系统自动分发：
Windows 走 ASan，Unix 走 TSan+UBSan。这导致 Linux/macOS 上的 ASan 覆盖
完全缺失——heap-use-after-free、栈溢出，以及自定义分配器（Arena、ColdData、
FixedFunction SBO、SmallVector）相关的内存错误，从未在 ASan 实现最强的
平台上被触发过。

改动：

* CMake：将 TASKFLOWLITE_ENABLE_SANITIZER (BOOL) 替换为
  TASKFLOWLITE_SANITIZER (STRING：OFF/ASAN/TSAN)，改为按值分发而非按 OS
  分发。对不支持的组合（MSVC + TSan）直接 FATAL_ERROR。保留旧布尔开关的
  deprecation shim 作为过渡。
  - Unix 上的 ASAN 顺带启用 UBSan，并加入 -fno-sanitize-recover=all，
    避免 UBSan 错误被 Catch2 静默吞掉。
  - TSAN 加入 -O1（TSan 官方推荐的最低优化等级，-O0 会漏掉部分 race）。

* CI 矩阵：6 → 8 个 job。为 Ubuntu 和 macOS 各新增一条 Debug+ASAN 配置，
  现在每个 Unix 平台都同时具备 ASan 和 TSan 双重视角。Windows 仍然仅
  保留 ASan（MSVC 无 TSan 支持）。

* CI：收紧 TSan suppressions 文件的写入条件——改为判断
  matrix.sanitizer == 'TSAN'，比之前的 `-n "$TSAN_OPTIONS"` 更清晰。
  新增 ASAN_OPTIONS / UBSAN_OPTIONS，设置 halt_on_error=1 和
  print_stacktrace=1，确保 sanitizer 报错能在 CI 日志中清晰呈现。

+++ taskflowlite/core/executor.hpp
@@ -431,7 +431,7 @@ inline void Executor::_spawn(std::size_t num_workers) {
         wr.m_adaptive_factor = 4;
         wr.m_max_steals = static_cast<std::uint32_t>(num_queues() * 2);
         wr.m_thread = std::thread([this, &wr]() noexcept {
-            wr.m_rng.seed(std::hash<std::thread::id>{}(std::this_thread::get_id()), num_queues());
+            wr.m_rng.seed(std::hash<std::thread::id>{}(std::this_thread::get_id()), static_cast<std::uint32_t>(num_queues()));
             m_handler.on_start(wr);
 
             Work* w = nullptr;
